requirements.txt
Django>=5.1,<6.0
gunicorn>=21
python-dotenv>=1.0
psycopg2-binary>=2.9

# Auth/Forms/CORS
django-allauth>=0.63
django-crispy-forms
django-cors-headers

# Utils
Pillow
requests
PyJWT
cryptography
python-decouple
whitenoise>=6.6

